+++ app/src/androidTest/java/au/com/shiftyjelly/pocketcasts/repositories/playlist/PlaylistManagerDsl.kt
@@ -498,8 +498,7 @@ class PlaylistManagerDsl : TestWatcher() {
             PlaylistPreviewForEpisode(
                 uuid = "playlist-id-$index",
                 title = "Playlist title $index",
-                episodeCount = 0,
-                hasEpisode = false,
+                episodeUuids = emptySet(),
             ),
         )
     }